requirements.txt
aws-cdk-lib==2.44.0
aws-cdk.aws-kinesisfirehose-alpha==2.44.0a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.44.0a0
constructs==10.1.103
black==22.8.0
isort==5.10.1
pyproject-flake8==5.0.4.post1
